requirements.txt
Flask>=2.0.0
google-cloud-vision>=3.0.0
google-cloud-translate>=3.6.0
openai>=1.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
